Added deleteing item per quantity

+++ src/actions.js
@@ -1,16 +1,37 @@
 import {
     ADD,
     CLEAR,
-    DELETE
+    DELETE,
+    ADD_EXISTING,
+    DELETE_EXISTING
 } from './types.js';
+import {store} from './store.js';
 
 export let idNum = 0;
 
-export const addToCart = (item) => {
+export const addToCart = (item, key) => {
+  console.log(store.getState());
   return {
     type: ADD,
     item,
-    id: ++idNum
+    id: ++idNum,
+    quantity: 1,
+    key
+  }
+}
+
+export const addExisting = (item, key) => {
+  let i = 0;
+  while (key !== store.getState().cart[i].key) {
+    i++;
+  }
+  let qty = store.getState().cart[i].quantity;
+  return {
+    type: ADD_EXISTING,
+    item,
+    key,
+    id: ++idNum,
+    quantity: ++qty
   }
 }
 
@@ -23,6 +44,19 @@ export const clearCartAct = () => {
 }
 
 export const delItem = (id) => {
+  let i = 0;
+  while (id !== store.getState().cart[i].id) {
+    i++;
+  }
+  let qty = store.getState().cart[i].quantity;
+  if (store.getState().cart[i].quantity > 1) {
+    return {
+      type: DELETE_EXISTING,
+      id: ++idNum,
+      i,
+      quantity: --qty
+    }
+  }
   return {
     type: DELETE,
     id
